+++ R/fitting.R
@@ -25,8 +25,7 @@
 #' @param method 'VARPRO', 'VARPRO_3P', 'TARQUIN' or 'LCMODEL'.
 #' @param w_ref water reference data for concentration scaling (optional).
 #' @param opts options to pass to the analysis method.
-#' @param parallel perform analysis in parallel (TRUE or FALSE).
-#' @param cores number of cores to use for parallel analysis.
+#' @param parallel perform analyses in parallel (TRUE or FALSE).
 #' @return MRS analysis object.
 #' @examples
 #' fname <- system.file("extdata","philips_spar_sdat_WS.SDAT",package="spant")
@@ -36,8 +35,8 @@
 #' fit_result <- fit_mrs(svs, basis)
 #' }
 #' @export
-fit_mrs <- function(metab, basis = NULL, method = 'VARPRO_3P', w_ref = NULL, opts = NULL, 
-                    parallel = FALSE, cores = 4) {
+fit_mrs <- function(metab, basis = NULL, method = 'VARPRO_3P', w_ref = NULL,
+                    opts = NULL,  parallel = FALSE) {
   
   # start the clock
   ptm <- proc.time()
@@ -80,7 +79,7 @@ fit_mrs <- function(metab, basis = NULL, method = 'VARPRO_3P', w_ref = NULL, opt
   }
   
   #if (parallel) {doSNOW::registerDoSNOW(makeCluster(cores, type = "SOCK"))}
-  # if (parallel) {doMC::registerDoMC(cores = cores)}
+  #if (parallel) {doMC::registerDoMC(cores = cores)}
   #doSNOW::registerDoSNOW(makeCluster(cores, type = "SOCK"))
  
   #if (parallel)  {
@@ -102,7 +101,8 @@ fit_mrs <- function(metab, basis = NULL, method = 'VARPRO_3P', w_ref = NULL, opt
     result_list <- plyr::alply(metab$data, c(2, 3, 4, 5, 6), varpro, 
                                acq_paras, basis, opts, 
                                .parallel = parallel, 
-                               .paropts = list(.inorder = TRUE),
+                               .paropts = list(.inorder = TRUE,
+                                               .packages = "spant"),
                                .progress = "text", .inform = FALSE)
     
   } else if (METHOD == "VARPRO_3P") {
@@ -116,7 +116,8 @@ fit_mrs <- function(metab, basis = NULL, method = 'VARPRO_3P', w_ref = NULL, opt
     result_list <- plyr::alply(metab$data, c(2, 3, 4, 5, 6),
                                varpro_3_para, acq_paras, basis, opts, 
                                .parallel = parallel, 
-                               .paropts = list(.inorder = TRUE),
+                               .paropts = list(.inorder = TRUE,
+                                               .packages = "spant"),
                                .progress = "text", .inform = FALSE)
     
   } else if (METHOD == "TARQUIN") {
@@ -150,7 +151,8 @@ fit_mrs <- function(metab, basis = NULL, method = 'VARPRO_3P', w_ref = NULL, opt
     result_list <- plyr::alply(metab$data, c(2, 3, 4, 5, 6), tarquin_fit, 
                                temp_mrs, basis_file, opts, 
                                .parallel = parallel, 
-                               .paropts = list(.inorder = TRUE),
+                               .paropts = list(.inorder = TRUE,
+                                               .packages = "spant"),
                                .progress = "text", .inform = FALSE)
     
                          #.paropts = list(.options.snow=snowopts),
@@ -188,7 +190,8 @@ fit_mrs <- function(metab, basis = NULL, method = 'VARPRO_3P', w_ref = NULL, opt
     result_list <- plyr::alply(metab$data, c(2,3,4,5,6), lcmodel_fit, 
                                temp_mrs, basis_file, opts,
                                .parallel = parallel, 
-                               .paropts = list(.inorder = TRUE),
+                               .paropts = list(.inorder = TRUE,
+                                               .packages = "spant"),
                                .progress = "text", .inform = FALSE)
   } else if (exists(method)) {
     message(paste("Using custom fit method :", method))
@@ -203,7 +206,8 @@ fit_mrs <- function(metab, basis = NULL, method = 'VARPRO_3P', w_ref = NULL, opt
     result_list <- plyr::alply(metab$data, c(2, 3, 4, 5, 6),
                                get(method), acq_paras, basis, opts, 
                                .parallel = parallel, 
-                               .paropts = list(.inorder = TRUE),
+                               .paropts = list(.inorder = TRUE,
+                                               .packages = "spant"),
                                .progress = "text", .inform = FALSE)
     
   } else {
@@ -237,14 +241,48 @@ fit_mrs <- function(metab, basis = NULL, method = 'VARPRO_3P', w_ref = NULL, opt
     }
   }
   
-  df_list_amps <- result_list[seq(from = 1, by = res_n,length.out = fit_num)]
-  amps = plyr::ldply(df_list_amps, data.frame)[-1]
-  df_list_crlbs <- result_list[seq(from = 2, by = res_n, length.out = fit_num)]
-  crlbs = plyr::ldply(df_list_crlbs, data.frame)[-1]
-  colnames(crlbs) <- paste(colnames(crlbs), ".sd", sep = "")
-  df_list_diags <- result_list[seq(from = 3, by = res_n, length.out = fit_num)]
-  diags = plyr::ldply(df_list_diags, data.frame)[-1]
   fits <- result_list[seq(from = 4, by = res_n, length.out = fit_num)]
+ 
+  # check for any masked voxels 
+  if (anyNA(fits)) {
+    na_res <- TRUE
+    na_res_vec <- is.na(fits)
+    na_res_vec_ind <- which(is.na(fits))
+    good_row <- which(!na_res_vec)[[1]]
+    rowN <- length(fits)
+  } else {
+    na_res <- FALSE
+  }
+  
+  df_list_amps  <- result_list[seq(from = 1, by = res_n, length.out = fit_num)]
+  df_list_crlbs <- result_list[seq(from = 2, by = res_n, length.out = fit_num)]
+  df_list_diags <- result_list[seq(from = 3, by = res_n, length.out = fit_num)]
+ 
+  if (na_res) {
+    amp_colnames <- colnames(df_list_amps[[good_row]])
+    amps <- as.data.frame(matrix(NA, rowN, length(amp_colnames)))
+    colnames(amps) <- amp_colnames
+    amps_not_na <- plyr::ldply(df_list_amps[!na_res_vec], data.frame)[-1]
+    amps[(!na_res_vec),] <- amps_not_na
+    
+    crlbs <- as.data.frame(matrix(NA, rowN, length(amp_colnames)))
+    crlbs_not_na <- plyr::ldply(df_list_crlbs[!na_res_vec], data.frame)[-1]
+    crlbs[(!na_res_vec),] <- crlbs_not_na
+    colnames(crlbs) <- paste(colnames(amps), ".sd", sep = "")
+    
+    diags_colnames <- colnames(df_list_diags[[good_row]])
+    diags <- as.data.frame(matrix(NA, rowN, length(diags_colnames)))
+    colnames(diags) <- diags_colnames
+    diags_not_na <- plyr::ldply(df_list_diags[!na_res_vec], data.frame)[-1]
+    diags[(!na_res_vec),] <- diags_not_na
+  } else {
+    amps <- plyr::ldply(df_list_amps, data.frame)[-1]
+  
+    crlbs <- plyr::ldply(df_list_crlbs, data.frame)[-1]
+    colnames(crlbs) <- paste(colnames(amps), ".sd", sep = "")
+  
+    diags <- plyr::ldply(df_list_diags, data.frame)[-1]
+  }
   
   res_tab <- cbind(labs, amps, crlbs, diags)
   res_tab[, 1:5] <- sapply(res_tab[, 1:5], as.numeric)