+++ NEWS.md
@@ -1,3 +1,10 @@
+# spant 0.17.0
+* Added a function to grid shift 2D MRSI data in the x/y direction.
+* Better plotting/fitting support for masking data by setting data points to NA.
+* Bug fix for interactive voxel selection position indicator.
+* Added mask_xy to mask voxels in a centered rectangular region.
+* Minor changes to improve parallel processing support.
+
 # spant 0.16.0
 * SNR is now defined as the max data point divided by the standard deviation of
 the noise (n.b. factor of two has been removed in-line with upcoming terminology